+++ libraries/itg3200filv05/ITG3200.h
@@ -156,14 +156,14 @@ public:
   bool isRawDataReady();
   // Gyro Sensors
   void readTemp(float *_Temp);
-  void readGyroRaw(int16_t  *_GyroXYZ);
-  void readGyroRaw(int16_t *_GyroX, int16_t *_GyroY, int16_t *_GyroZ);
+  void readGyroRaw(int  *_GyroXYZ);
+  void readGyroRaw(int *_GyroX, int *_GyroY, int *_GyroZ);
   void setRevPolarity(bool _Xpol, bool _Ypol, bool _Zpol);	// true = Reversed  false = default
   void setGains(float _Xgain, float _Ygain, float _Zgain);
   void setOffsets(int _Xoffset, int _Yoffset, int _Zoffset);
   void zeroCalibrate(unsigned int totSamples, unsigned int sampleDelayMS);	// assuming gyroscope is stationary (updates XYZ offsets for zero)
-  void readGyroRawCal(int16_t *_GyroX, int16_t *_GyroY, int16_t *_GyroZ);
-  void readGyroRawCal(int16_t *_GyroXYZ);
+  void readGyroRawCal(int *_GyroX, int *_GyroY, int *_GyroZ);
+  void readGyroRawCal(int *_GyroXYZ);
   void readGyro(float *_GyroXYZ); // includes gain and offset
   void readGyro(float *_GyroX, float *_GyroY, float *_GyroZ); // includes gain and offset    
   // Power management
